merchant

+++ app/controllers/upload_file_controller.rb
@@ -28,13 +28,31 @@ class UploadFileController < ApplicationController
 				flash[:notice],flash[:error]=reconciliation.valid_reconciliation_by_country("de",filename)
 
 				File.delete(filename)
+			elsif params['file_type']=="MERCHANT_DE_AT_CASH_IN_BANK"
+				filename=write_file(params['file'],"xlsx")
+				reconciliation=ReconciliationSofort.new()
+				flash[:notice],flash[:error]=reconciliation.merchant_cash_in_proc("de_at",filename)
+
+				File.delete(filename)
+			elsif params['file_type']=="MERCHANT_EN_CASH_IN_BANK"
+				filename=write_file(params['file'],"xlsx")
+				reconciliation=ReconciliationSofort.new()
+				flash[:notice],flash[:error]=reconciliation.merchant_cash_in_proc("en",filename)
+
+				File.delete(filename)
+			elsif params['file_type']=="MERCHANT_NL_CASH_IN_BANK"
+				filename=write_file(params['file'],"xlsx")
+				reconciliation=ReconciliationSofort.new()
+				flash[:notice],flash[:error]=reconciliation.merchant_cash_in_proc("nl",filename)
+
+				File.delete(filename)
 			else
 				flash[:notice]="未定义的上传业务类型,请重新选择"
 			end
 		rescue=>e
 			logger.info(e.message)
 			flash[:notice],flash[:error]="处理文件失败,请重试",e.message
-			File.delete(filename)
+			File.delete(filename) if File.exists?(filename)
 		end
 
 		render upload_file_index_path
